requirements.txt
flask>=2.0.0
flask-cors>=3.0.0
opencv-python>=4.5.0
numpy>=1.19.0
pyyaml>=5.4.0
openpose-python
pytest>=6.0.0
flake8>=3.9.0
# TODO: 添加版本约束
# TODO: 添加开发依赖
# TODO: 添加可选依赖
# TODO: 分离生产和开发依赖
  